requirements.txt
alembic==1.7.5
attrs==21.2.0
black==21.12b0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==6.2
dnspython==2.1.0
email-validator==1.1.3
flake8==4.0.1
Flask==2.1.3
Flask-DebugToolbar==0.11.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.1.6
MarkupSafe==2.1.5
mccabe==0.6.1
mypy==0.920
mypy-extensions==0.4.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.2
requests==2.32.3
SQLAlchemy==1.4.28
toml==0.10.2
tomli==1.2.3
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.0.1
urllib3==2.2.2
Werkzeug==2.3.1
WTForms==3.0.0
